+++ integration/hsm/hsm_test.go
@@ -453,10 +453,10 @@ func TestHSMMigrate(t *testing.T) {
 	// start a dual auth non-hsm cluster
 	log.Debug("TestHSMMigrate: Starting auth server 1")
 	auth1Config := newHSMAuthConfig(t, storageConfig, log)
-	auth1Config.Auth.KeyStore = keystore.Config{}
+	auth1Config.Auth.KeyStore = servicecfg.KeystoreConfig{}
 	auth1 := newTeleportService(t, auth1Config, "auth1")
 	auth2Config := newHSMAuthConfig(t, storageConfig, log)
-	auth2Config.Auth.KeyStore = keystore.Config{}
+	auth2Config.Auth.KeyStore = servicecfg.KeystoreConfig{}
 	auth2 := newTeleportService(t, auth2Config, "auth2")
 	require.NoError(t, auth1.start(ctx))
 	require.NoError(t, auth2.start(ctx))
