+++ integration/proxy/proxy_tunnel_strategy_test.go
@@ -36,7 +36,7 @@ import (
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/auth/testauthority"
-	"github.com/gravitational/teleport/lib/cloud"
+	"github.com/gravitational/teleport/lib/cloud/imds"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/modules"
 	"github.com/gravitational/teleport/lib/service"
@@ -334,7 +334,7 @@ func (p *proxyTunnelStrategy) makeProxy(t *testing.T) {
 	conf.SetToken("token")
 	conf.DataDir = t.TempDir()
 	conf.Log = proxy.Log
-	conf.InstanceMetadataClient = cloud.NewDisabledIMDSClient()
+	conf.InstanceMetadataClient = imds.NewDisabledIMDSClient()
 
 	conf.Auth.Enabled = false
 	conf.SSH.Enabled = false
@@ -379,7 +379,7 @@ func (p *proxyTunnelStrategy) makeNode(t *testing.T) {
 	conf.SetToken("token")
 	conf.DataDir = t.TempDir()
 	conf.Log = node.Log
-	conf.InstanceMetadataClient = cloud.NewDisabledIMDSClient()
+	conf.InstanceMetadataClient = imds.NewDisabledIMDSClient()
 
 	conf.Auth.Enabled = false
 	conf.Proxy.Enabled = false
@@ -424,7 +424,7 @@ func (p *proxyTunnelStrategy) makeDatabase(t *testing.T) {
 	conf.SetToken("token")
 	conf.DataDir = t.TempDir()
 	conf.Log = db.Log
-	conf.InstanceMetadataClient = cloud.NewDisabledIMDSClient()
+	conf.InstanceMetadataClient = imds.NewDisabledIMDSClient()
 
 	conf.Auth.Enabled = false
 	conf.Proxy.Enabled = false
