+++ lib/cloud/imds/azure/imds_test.go
