+++ lib/cloud/imds/gcp/imds.go
