+++ lib/labels/cloud.go
@@ -26,7 +26,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib/cloud"
+	"github.com/gravitational/teleport/lib/cloud/imds"
 )
 
 const (
@@ -47,7 +47,7 @@ const (
 
 // CloudConfig is the configuration for a cloud label service.
 type CloudConfig struct {
-	Client               cloud.InstanceMetadata
+	Client               imds.Client
 	Clock                clockwork.Clock
 	Log                  logrus.FieldLogger
 	namespace            string
