+++ lib/service/service.go
@@ -82,6 +82,7 @@ import (
 	"github.com/gravitational/teleport/lib/auth/accesspoint"
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	"github.com/gravitational/teleport/lib/auth/keygen"
+	"github.com/gravitational/teleport/lib/auth/keystore"
 	"github.com/gravitational/teleport/lib/auth/native"
 	"github.com/gravitational/teleport/lib/auth/state"
 	"github.com/gravitational/teleport/lib/authz"
@@ -1817,6 +1818,23 @@ func (process *TeleportProcess) initAuthService() error {
 	if err != nil {
 		return trace.Wrap(err)
 	}
+
+	keystoreConfig := keystore.Config{
+		PKCS11: keystore.PKCS11Config{
+			Path:       cfg.Auth.KeyStore.PKCS11.Path,
+			SlotNumber: cfg.Auth.KeyStore.PKCS11.SlotNumber,
+			TokenLabel: cfg.Auth.KeyStore.PKCS11.TokenLabel,
+			Pin:        cfg.Auth.KeyStore.PKCS11.Pin,
+			HostUUID:   cfg.Auth.KeyStore.PKCS11.HostUUID,
+		},
+		GCPKMS: keystore.GCPKMSConfig{
+			KeyRing:         cfg.Auth.KeyStore.GCPKMS.KeyRing,
+			ProtectionLevel: cfg.Auth.KeyStore.GCPKMS.ProtectionLevel,
+			HostUUID:        cfg.Auth.KeyStore.GCPKMS.HostUUID,
+		},
+		Logger: process.log,
+	}
+
 	// first, create the AuthServer
 	authServer, err := auth.Init(
 		process.ExitContext(),
@@ -1849,7 +1867,7 @@ func (process *TeleportProcess) initAuthService() error {
 			OIDCConnectors:          cfg.OIDCConnectors,
 			AuditLog:                process.auditLog,
 			CipherSuites:            cfg.CipherSuites,
-			KeyStoreConfig:          cfg.Auth.KeyStore,
+			KeyStoreConfig:          keystoreConfig,
 			Emitter:                 checkingEmitter,
 			Streamer:                events.NewReportingStreamer(streamer, process.Config.Testing.UploadEventsC),
 			TraceClient:             traceClt,
