+++ lib/srv/discovery/fetchers/aks.go
@@ -88,7 +88,7 @@ func (a *aksFetcher) Get(ctx context.Context) (types.ResourcesWithLabels, error)
 			a.Log.Debugf("Cluster region %q does not match with allowed values.", cluster.Location)
 			continue
 		}
-		kubeCluster, err := services.NewKubeClusterFromAzureAKS(cluster)
+		kubeCluster, err := common.NewKubeClusterFromAzureAKS(cluster)
 		if err != nil {
 			a.Log.WithError(err).Warn("Unable to create Kubernetes cluster from azure.AKSCluster.")
 			continue
