+++ lib/srv/discovery/fetchers/kube_services_test.go
@@ -268,7 +268,10 @@ func TestKubeAppFetcher_Get(t *testing.T) {
 	}
 
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.desc, func(t *testing.T) {
+			t.Parallel()
+
 			var objects []runtime.Object
 			for _, s := range tt.services {
 				objects = append(objects, s)
@@ -293,7 +296,6 @@ func TestKubeAppFetcher_Get(t *testing.T) {
 			})
 			require.Empty(t, cmp.Diff(tt.expected.AsResources(), result))
 		})
-
 	}
 }
